adjusted font links
fixed font awesome and added grass effect to header
I have a feeling font awesome will fail again, if so I commented the
URL to the CDN, it would be nice if I could be free of CDNs though.
Also made a copy of the link for the google fonts in case that messes
things up.